third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-document-suborigin-expected.txt

CONSOLE ERROR: Unrecognized Content-Security-Policy directive 'suborigin'.

CONSOLE ERROR: line 2608: Uncaught assert_equals: document.suborigin should exist when page has a suborigin. expected (string) "foobar" but got (undefined) undefined
This is a testharness.js-based test.
FAIL Validate presence and value of document.suborigin Uncaught assert_equals: document.suborigin should exist when page has a suborigin. expected (string) "foobar" but got (undefined) undefined
Harness: the test ran to completion.
